backend/requirements.txt
# backend/requirements.txt

fastapi
numpy
transformers
pillow
torch 
python-multipart
uvicorn